add: module support for /popup/utils.js
