+++ popup/popup.html
@@ -21,5 +21,5 @@
     <button id="monitor">Monitor</button>
 </div>
 </body>
-<script src="popup.js"></script>
+<script type="module" src="popup.js"></script>
 </html>
