+++ popup/popup.js
@@ -1,3 +1,7 @@
+import {
+    addClickListener, addBlurListener, getInputValue, setInputValue, resetInputStyles, highlightInvalidField, isValidRegex, isValidNumber, updateButtonState
+} from "./utils.js";
+
 document.addEventListener("DOMContentLoaded", () => {
     initializeEventListeners();
     loadSettings();
@@ -13,12 +17,14 @@ function initializeEventListeners() {
     addBlurListener("openingDelay", saveSettings);
 }
 
-function addClickListener(elementId, handler) {
-    document.getElementById(elementId).addEventListener("click", handler);
-}
-
-function addBlurListener(elementId, handler) {
-    document.getElementById(elementId).addEventListener("blur", handler);
+function toggleSetting(settingKey, buttonId) {
+    chrome.storage.local.get([settingKey], (result) => {
+        const newState = !result[settingKey];
+        chrome.storage.local.set({ [settingKey]: newState }, () => {
+            updateButtonState(buttonId, newState);
+            console.log(`${settingKey} toggled:`, newState);
+        });
+    });
 }
 
 function loadSettings() {
@@ -34,10 +40,6 @@ function loadSettings() {
     });
 }
 
-function setInputValue(elementId, value, defaultValue = "") {
-    document.getElementById(elementId).value = value !== undefined ? value : defaultValue;
-}
-
 function saveSettings() {
     const settings = {
         channelUrl: getInputValue("channelUrl"),
@@ -45,7 +47,7 @@ function saveSettings() {
         openingDelay: getInputValue("openingDelay")
     };
 
-    resetInputStyles();
+    resetInputStyles(["channelUrl", "regexFilter", "openingDelay"]);
 
     if (!settings.channelUrl) {
         highlightInvalidField("channelUrl");
@@ -65,31 +67,6 @@ function saveSettings() {
     chrome.storage.local.set(settings, () => console.log("Settings saved"));
 }
 
-function getInputValue(elementId) {
-    return document.getElementById(elementId).value;
-}
-
-function resetInputStyles() {
-    ["channelUrl", "regexFilter", "openingDelay"].forEach(id => document.getElementById(id).style.backgroundColor = "");
-}
-
-function highlightInvalidField(elementId) {
-    document.getElementById(elementId).style.backgroundColor = "rgba(190,25,43,0.2)";
-}
-
-function isValidRegex(value) {
-    try {
-        new RegExp(value);
-        return true;
-    } catch (e) {
-        return false;
-    }
-}
-
-function isValidNumber(value) {
-    return !isNaN(value) && Number(value) >= 0;
-}
-
 function handleMonitorClick() {
     const channelUrl = getInputValue("channelUrl");
 
@@ -106,20 +83,4 @@ function handleMonitorClick() {
         window.close();
         chrome.tabs.sendMessage(currentTab.id, { type: "openDiscord", url: channelUrl });
     });
-}
-
-function toggleSetting(settingKey, buttonId) {
-    chrome.storage.local.get([settingKey], (result) => {
-        const newState = !result[settingKey];
-        chrome.storage.local.set({ [settingKey]: newState }, () => {
-            updateButtonState(buttonId, newState);
-            console.log(`${settingKey} toggled:`, newState);
-        });
-    });
-}
-
-function updateButtonState(buttonId, isEnabled) {
-    const button = document.getElementById(buttonId);
-    button.classList.toggle("enabled", isEnabled);
-    button.classList.toggle("disabled", !isEnabled);
 }
